requirements.txt
Flask
flask_restful
flask_cors
flask_marshmallow
marshmallow-sqlalchemy
PyJWT
Flask-SQLAlchemy
datetime
flask_jwt_extended